Completed Rank Model
Completed the team rank model.  Added functions that determine if a specific feature from the model is zero (i.e., determine if feature is part of the subset).

Specifically, runs a normality test on the feature and if it is Gaussian, runs a one-sample t-test.  If it is not, runs a Wilcoxon test.  Should only be run if rank model utilizes a Bayesian feature selection rank model.

Also, added dictionary outputs that maps the column names to the feature names for the rank model (bayes_TeamRankModel) and the team model (bayes_TeamStat).